requirements.txt
# Основные зависимости для YOLOv8 и компьютерного зрения
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0

# Дополнительные зависимости для работы с данными
pyyaml>=6.0
pandas>=1.5.0

# Для экспорта в ONNX
onnx>=1.14.0
onnxruntime>=1.15.0
